Some manual fixes to instrument.js
It turned out that the minimizer for sending files to the frontend
inserts semicolons at the end of lines. The result is that when an if
else statement is formated in a certain way it will break the
  application.

E.g. :

if (foo) { doSomething() }
else { doSomethingElse() }

will be minified to something like this:

if(foo){doSomething()};else{doSomethingElse()}

Note the semicolon *before* else here. This results in a syntax error.

The `fixmyjs` script I used so far adds curly brackets around bodies of
if/else/while/for blocks. This sometimes triggers above bug. In this
  commit I manually inserted curly brackets at places where `fixmyjs`
  will mess this up, such that I can later run `fixmyjs` to improve the
  rest.